+++ src/renderer/src/components/ui/modal/stacked/provider.tsx
@@ -4,7 +4,7 @@ import type { FC, PropsWithChildren } from "react"
 
 import { modalStackAtom } from "./atom"
 import { MODAL_STACK_Z_INDEX } from "./constants"
-import { useDismissAllWhenRouterChange } from "./hooks"
+// import { useDismissAllWhenRouterChange } from "./hooks"
 import { ModalInternal } from "./modal"
 import { ModalOverlay } from "./overlay"
 
@@ -18,7 +18,8 @@ export const ModalStackProvider: FC<PropsWithChildren> = ({ children }) => (
 const ModalStack = () => {
   const stack = useAtomValue(modalStackAtom)
 
-  useDismissAllWhenRouterChange()
+  // Vite HMR issue
+  // useDismissAllWhenRouterChange()
 
   return (
     <AnimatePresence mode="popLayout">
