+++ src/renderer/src/styles/colors.css
@@ -14,6 +14,9 @@
 
   --fo-item-active: theme(colors.native.active/0.4);
   --fo-item-hover: theme(colors.native.active/0.2);
+
+  --fo-modal-background: theme(colors.zinc.50/0.8);
+  --fo-modal-background-opaque: theme(colors.zinc.50);
 }
 
 [data-theme="dark"] {
@@ -31,4 +34,7 @@
 
   --fo-text-primary: 0, 0%, 80%;
   --fo-background: theme(colors.background);
+
+  --fo-modal-background: theme(colors.neutral.900/0.8);
+  --fo-modal-background-opaque: theme(colors.neutral.900);
 }
